fix: reprogram standard library for env api
