+++ lib/values/symbol.ts
@@ -1,38 +1,33 @@
-interface Symbol {
-    valueOf(): symbol;
-    constructor: SymbolConstructor;
-}
-interface SymbolConstructor {
-    (val?: any): symbol;
-    prototype: Symbol;
-    for(key: string): symbol;
-    keyFor(sym: symbol): string;
-    readonly typeName: unique symbol;
-}
+define("values/symbol", () => {
+    var Symbol = env.global.Symbol = function(this: any, val?: string) {
+        if (this !== undefined && this !== null) throw new TypeError("Symbol may not be called with 'new'.");
+        if (typeof val !== 'string' && val !== undefined) throw new TypeError('val must be a string or undefined.');
+        return env.internals.symbol(val, true);
+    } as SymbolConstructor;
 
-declare var Symbol: SymbolConstructor;
+    Symbol.prototype = env.internals.symbolProto;
+    setConstr(Symbol.prototype, Symbol, env);
+    (Symbol as any).typeName = Symbol("Symbol.name");
+    (Symbol as any).replace = Symbol('Symbol.replace');
+    (Symbol as any).match = Symbol('Symbol.match');
+    (Symbol as any).matchAll = Symbol('Symbol.matchAll');
+    (Symbol as any).split = Symbol('Symbol.split');
+    (Symbol as any).search = Symbol('Symbol.search');
+    (Symbol as any).iterator = Symbol('Symbol.iterator');
+    (Symbol as any).asyncIterator = Symbol('Symbol.asyncIterator');
 
-gt.Symbol = function(this: any, val?: string) {
-    if (this !== undefined && this !== null) throw new TypeError("Symbol may not be called with 'new'.");
-    if (typeof val !== 'string' && val !== undefined) throw new TypeError('val must be a string or undefined.');
-    return internals.symbol(val, true);
-} as SymbolConstructor;
+    setProps(Symbol, env, {
+        for(key) {
+            if (typeof key !== 'string' && key !== undefined) throw new TypeError('key must be a string or undefined.');
+            return env.internals.symbol(key, false);
+        },
+        keyFor(sym) {
+            if (typeof sym !== 'symbol') throw new TypeError('sym must be a symbol.');
+            return env.internals.symStr(sym);
+        },
+        typeName: Symbol('Symbol.name') as any,
+    });
 
-Symbol.prototype = internals.symbolProto;
-setConstr(Symbol.prototype, Symbol);
-(Symbol as any).typeName = Symbol("Symbol.name");
-
-setProps(Symbol, {
-    for(key) {
-        if (typeof key !== 'string' && key !== undefined) throw new TypeError('key must be a string or undefined.');
-        return internals.symbol(key, false);
-    },
-    keyFor(sym) {
-        if (typeof sym !== 'symbol') throw new TypeError('sym must be a symbol.');
-        return internals.symStr(sym);
-    },
-    typeName: Symbol('Symbol.name') as any,
+    env.global.Object.defineProperty(Object.prototype, Symbol.typeName, { value: 'Object' });
+    env.global.Object.defineProperty(env.global, Symbol.typeName, { value: 'Window' });
 });
-
-Object.defineProperty(Object.prototype, Symbol.typeName, { value: 'Object' });
-Object.defineProperty(gt, Symbol.typeName, { value: 'Window' });
