+++ tsconfig.json
@@ -1,12 +1,29 @@
 {
-    "include": [ "lib/**/*.ts" ],
+    "files": [
+        "lib/lib.d.ts",
+        "lib/modules.ts",
+        "lib/utils.ts",
+        "lib/values/object.ts",
+        "lib/values/symbol.ts",
+        "lib/values/function.ts",
+        "lib/values/errors.ts",
+        "lib/values/string.ts",
+        "lib/values/number.ts",
+        "lib/values/boolean.ts",
+        "lib/values/array.ts",
+        "lib/promise.ts",
+        "lib/map.ts",
+        "lib/set.ts",
+        "lib/regex.ts",
+        "lib/core.ts"
+    ],
     "compilerOptions": {
-        "outDir": "bin/me/topchetoeu/jscript/js",
-        "declarationDir": "bin/me/topchetoeu/jscript/dts",
+        "outFile": "bin/me/topchetoeu/jscript/js/core.js",
+        // "declarationDir": "",
+        // "declarationDir": "bin/me/topchetoeu/jscript/dts",
         "target": "ES5",
         "lib": [],
-        "module": "CommonJS",
-        "declaration": true,
+        "module": "None",
         "stripInternal": true,
         "downlevelIteration": true,
         "esModuleInterop": true,
